[SPARK-11265][YARN] YarnClient can't get tokens to talk to Hive in a secure…
Backport to branch-1.5 of SPARK-11265 patch.

The sole change is in `Client`, where there's no longer a probe to see if the token request is enabled; it always happens. This means that provided there's hive.jar on the classpath, this will attempt to get the token —and if that fails for any reason other than CNFE, the client launch will fail

Author: Steve Loughran <stevel@hortonworks.com>

Closes #9466 from steveloughran/stevel/patches/SPARK-11265-on-branch-1.5.

(cherry picked from commit d31b312fc337c27eba346d3682fdd2efe7ac9a0f)